Support aspect ratio for Gemini image generation.
