+++ src/Models/GoogleImageGenerationModel.php
@@ -7,7 +7,6 @@ namespace WordPress\GoogleAiProvider\Models;
 use WordPress\AiClient\Common\Exception\InvalidArgumentException;
 use WordPress\AiClient\Common\Exception\RuntimeException;
 use WordPress\AiClient\Files\DTO\File;
-use WordPress\AiClient\Files\Enums\MediaOrientationEnum;
 use WordPress\AiClient\Messages\DTO\Message;
 use WordPress\AiClient\Messages\DTO\MessagePart;
 use WordPress\AiClient\Messages\Enums\MessageRoleEnum;
@@ -53,6 +52,8 @@ use WordPress\GoogleAiProvider\Provider\GoogleProvider;
  */
 class GoogleImageGenerationModel extends AbstractApiBasedModel implements ImageGenerationModelInterface
 {
+    use WithAspectRatioTrait;
+
     /**
      * {@inheritDoc}
      *
@@ -203,34 +204,6 @@ class GoogleImageGenerationModel extends AbstractApiBasedModel implements ImageG
         }
 
         return $text;
-    }
-
-    /**
-     * Prepares the aspect ratio parameter for the API request.
-     *
-     * @since 1.0.0
-     *
-     * @param MediaOrientationEnum|null $orientation The desired media orientation.
-     * @param string|null $aspectRatio The desired media aspect ratio.
-     * @return string The prepared aspect ratio parameter.
-     */
-    protected function prepareAspectRatioParam(?MediaOrientationEnum $orientation, ?string $aspectRatio): string
-    {
-        // Use aspect ratio if set, as it is more specific.
-        if ($aspectRatio !== null) {
-            return $aspectRatio;
-        }
-
-        // This should always have a value, as the method is only called if at least one or the other is set.
-        if ($orientation !== null) {
-            if ($orientation->isLandscape()) {
-                return '16:9';
-            }
-            if ($orientation->isPortrait()) {
-                return '9:16';
-            }
-        }
-        return '1:1';
     }
 
     /**
